refactors and styling

+++ client/src/main/java/ru/mrbedrockpy/craftengine/world/TickSystem.java
@@ -40,6 +40,6 @@ public class TickSystem {
     }
 
     public double partialTick() {
-        return accumulator / tickTime;
+        return ((double) accumulator) / ((double) tickTime);
     }
 }
